Marshal XML To Create Bulk Jobs
Marshal JobInfo structs to create Bulk API jobs rather than generating
XML using strings.

Move Bulk API functions into a new bulk.go file.
