+++ lib/force.go
@@ -160,43 +160,6 @@ type QueryOptions struct {
 	IsTooling bool
 }
 
-type BatchResult struct {
-	Results []Result
-}
-
-type BatchInfo struct {
-	Id                     string `xml:"id"`
-	JobId                  string `xml:"jobId"`
-	State                  string `xml:"state"`
-	CreatedDate            string `xml:"createdDate"`
-	SystemModstamp         string `xml:"systemModstamp"`
-	NumberRecordsProcessed int    `xml:"numberRecordsProcessed"`
-}
-
-type JobInfo struct {
-	Id                      string `xml:"id"`
-	Operation               string `xml:"operation"`
-	Object                  string `xml:"object"`
-	CreatedById             string `xml:"createdById"`
-	CreatedDate             string `xml:"createdDate"`
-	SystemModStamp          string `xml:"systemModstamp"`
-	State                   string `xml:"state"`
-	ContentType             string `xml:"contentType"`
-	ConcurrencyMode         string `xml:"concurrencyMode"`
-	NumberBatchesQueued     int    `xml:"numberBatchesQueued"`
-	NumberBatchesInProgress int    `xml:"numberBatchesInProgress"`
-	NumberBatchesCompleted  int    `xml:"numberBatchesCompleted"`
-	NumberBatchesFailed     int    `xml:"numberBatchesFailed"`
-	NumberBatchesTotal      int    `xml:"numberBatchesTotal"`
-	NumberRecordsProcessed  int    `xml:"numberRecordsProcessed"`
-	NumberRetries           int    `xml:"numberRetries"`
-	ApiVersion              string `xml:"apiVersion"`
-	NumberRecordsFailed     int    `xml:"numberRecordsFailed"`
-	TotalProcessingTime     int    `xml:"totalProcessingTime"`
-	ApiActiveProcessingTime int    `xml:"apiActiveProcessingTime"`
-	ApexProcessingTime      int    `xml:"apexProcessingTime"`
-}
-
 type AuraDefinitionBundleResult struct {
 	Done           bool
 	Records        []ForceRecord
@@ -947,139 +910,6 @@ func (f *Force) CreateRecord(sobject string, attrs map[string]string) (id string
 	var result ForceCreateRecordResult
 	json.Unmarshal(body, &result)
 	id = result.Id
-	return
-}
-
-func (f *Force) CreateBulkJob(xmlbody string) (result JobInfo, err error) {
-	url := fmt.Sprintf("%s/services/async/%s/job", f.Credentials.InstanceUrl, apiVersionNumber)
-	body, err := f.httpPostXML(url, xmlbody)
-	xml.Unmarshal(body, &result)
-	if len(result.Id) == 0 {
-		var fault LoginFault
-		xml.Unmarshal(body, &fault)
-		err = errors.New(fmt.Sprintf("%s: %s", fault.ExceptionCode, fault.ExceptionMessage))
-	}
-	return
-}
-
-func (f *Force) CloseBulkJob(jobId string, xmlbody string) (result JobInfo, err error) {
-	url := fmt.Sprintf("%s/services/async/%s/job/%s", f.Credentials.InstanceUrl, apiVersionNumber, jobId)
-	body, err := f.httpPostXML(url, xmlbody)
-	xml.Unmarshal(body, &result)
-	if len(result.Id) == 0 {
-		var fault LoginFault
-		xml.Unmarshal(body, &fault)
-		err = errors.New(fmt.Sprintf("%s: %s", fault.ExceptionCode, fault.ExceptionMessage))
-	}
-	return
-}
-
-func (f *Force) GetBulkJobs() (result []JobInfo, err error) {
-	url := fmt.Sprintf("%s/services/async/%s/jobs", f.Credentials.InstanceUrl, apiVersionNumber)
-	body, err := f.httpGetBulk(url)
-	xml.Unmarshal(body, &result)
-	if len(result[0].Id) == 0 {
-		var fault LoginFault
-		xml.Unmarshal(body, &fault)
-		err = errors.New(fmt.Sprintf("%s: %s", fault.ExceptionCode, fault.ExceptionMessage))
-	}
-	return
-}
-
-func (f *Force) BulkQuery(soql string, jobId string, contettype string) (result BatchInfo, err error) {
-	url := fmt.Sprintf("%s/services/async/%s/job/%s/batch", f.Credentials.InstanceUrl, apiVersionNumber, jobId)
-	var body []byte
-
-	if contettype == "CSV" {
-		body, err = f.httpPostCSV(url, soql)
-		xml.Unmarshal(body, &result)
-	} else {
-		body, err = f.httpPostXML(url, soql)
-		xml.Unmarshal(body, &result)
-	}
-	if len(result.Id) == 0 {
-		var fault LoginFault
-		xml.Unmarshal(body, &fault)
-		err = errors.New(fmt.Sprintf("%s: %s", fault.ExceptionCode, fault.ExceptionMessage))
-	}
-	return
-}
-
-func (f *Force) AddBatchToJob(xmlbody string, jobId string) (result BatchInfo, err error) {
-	url := fmt.Sprintf("%s/services/async/%s/job/%s/batch", f.Credentials.InstanceUrl, apiVersionNumber, jobId)
-	body, err := f.httpPostCSV(url, xmlbody)
-	xml.Unmarshal(body, &result)
-	if len(result.Id) == 0 {
-		var fault LoginFault
-		xml.Unmarshal(body, &fault)
-		err = errors.New(fmt.Sprintf("%s: %s", fault.ExceptionCode, fault.ExceptionMessage))
-	}
-	return
-}
-
-func (f *Force) GetBatchInfo(jobId string, batchId string) (result BatchInfo, err error) {
-	url := fmt.Sprintf("%s/services/async/%s/job/%s/batch/%s", f.Credentials.InstanceUrl, apiVersionNumber, jobId, batchId)
-	body, err := f.httpGetBulk(url)
-	xml.Unmarshal(body, &result)
-	if len(result.Id) == 0 {
-		var fault LoginFault
-		xml.Unmarshal(body, &fault)
-		err = errors.New(fmt.Sprintf("%s: %s", fault.ExceptionCode, fault.ExceptionMessage))
-	}
-	return
-}
-
-func (f *Force) GetBatches(jobId string) (result []BatchInfo, err error) {
-	url := fmt.Sprintf("%s/services/async/%s/job/%s/batch", f.Credentials.InstanceUrl, apiVersionNumber, jobId)
-	body, err := f.httpGetBulk(url)
-
-	var batchInfoList struct {
-		BatchInfos []BatchInfo `xml:"batchInfo"`
-	}
-
-	xml.Unmarshal(body, &batchInfoList)
-	result = batchInfoList.BatchInfos
-	if len(result) == 0 {
-		var fault LoginFault
-		xml.Unmarshal(body, &fault)
-		err = errors.New(fmt.Sprintf("%s: %s", fault.ExceptionCode, fault.ExceptionMessage))
-	}
-	return
-}
-
-func (f *Force) GetJobInfo(jobId string) (result JobInfo, err error) {
-	url := fmt.Sprintf("%s/services/async/%s/job/%s", f.Credentials.InstanceUrl, apiVersionNumber, jobId)
-	body, err := f.httpGetBulk(url)
-	xml.Unmarshal(body, &result)
-	if len(result.Id) == 0 {
-		var fault LoginFault
-		xml.Unmarshal(body, &fault)
-		err = errors.New(fmt.Sprintf("%s: %s", fault.ExceptionCode, fault.ExceptionMessage))
-	}
-	return
-}
-
-func (f *Force) RetrieveBulkQuery(jobId string, batchId string) (result []byte, err error) {
-	url := fmt.Sprintf("%s/services/async/%s/job/%s/batch/%s/result", f.Credentials.InstanceUrl, apiVersionNumber, jobId, batchId)
-	result, err = f.httpGetBulk(url)
-	return
-}
-
-func (f *Force) RetrieveBulkQueryResults(jobId string, batchId string, resultId string) (result []byte, err error) {
-	url := fmt.Sprintf("%s/services/async/%s/job/%s/batch/%s/result/%s", f.Credentials.InstanceUrl, apiVersionNumber, jobId, batchId, resultId)
-	result, err = f.httpGetBulk(url)
-	return
-}
-
-func (f *Force) RetrieveBulkBatchResults(jobId string, batchId string) (results BatchResult, err error) {
-	url := fmt.Sprintf("%s/services/async/%s/job/%s/batch/%s/result", f.Credentials.InstanceUrl, apiVersionNumber, jobId, batchId)
-	result, err := f.httpGetBulk(url)
-	if len(result) == 0 {
-		var fault LoginFault
-		xml.Unmarshal(result, &fault)
-		err = errors.New(fmt.Sprintf("%s: %s", fault.ExceptionCode, fault.ExceptionMessage))
-	}
-	//	sreader = Reader.NewReader(result);
 	return
 }
